Little refactoring;
1. Was deleted class Generic
2. It was replaced by GenericType for TypeDefBlock
